refactor(rule): consolidate credit absence tests and enhance processor logic with new test cases
